Struct stm32l0::stm32l0x1::rcc::apb1smenr::W

pub struct W { /* fields omitted */ }

Value to write to the register

Methods

impl W

pub fn reset_value() -> W

Reset value of the register

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register

pub fn lptim1smen(&mut self) -> _LPTIM1SMENW

Bit 31 - Low power timer clock enable during sleep mode bit

pub fn pwrsmen(&mut self) -> _PWRSMENW

Bit 28 - Power interface clock enable during sleep mode bit

pub fn crssmen(&mut self) -> _CRSSMENW

Bit 27 - Clock recovery system clock enable during sleep mode bit

pub fn i2c2smen(&mut self) -> _I2C2SMENW

Bit 22 - I2C2 clock enable during sleep mode bit

pub fn i2c1smen(&mut self) -> _I2C1SMENW

Bit 21 - I2C1 clock enable during sleep mode bit

pub fn lpuart1smen(&mut self) -> _LPUART1SMENW

Bit 18 - LPUART1 clock enable during sleep mode bit

pub fn usart2smen(&mut self) -> _USART2SMENW

Bit 17 - UART2 clock enable during sleep mode bit

pub fn spi2smen(&mut self) -> _SPI2SMENW

Bit 14 - SPI2 clock enable during sleep mode bit

pub fn wwdgsmen(&mut self) -> _WWDGSMENW

Bit 11 - Window watchdog clock enable during sleep mode bit

pub fn tim6smen(&mut self) -> _TIM6SMENW

Bit 4 - Timer 6 clock enable during sleep mode bit

pub fn tim2smen(&mut self) -> _TIM2SMENW

Bit 0 - Timer2 clock enable during sleep mode bit

pub fn tim3smen(&mut self) -> _TIM3SMENW

Bit 1 - Timer 3 clock enable during sleep mode bit

pub fn tim7smen(&mut self) -> _TIM7SMENW

Bit 5 - Timer 7 clock enable during sleep mode bit

pub fn usart4smen(&mut self) -> _USART4SMENW

Bit 19 - USART4 clock enabe during sleep mode bit

pub fn usart5smen(&mut self) -> _USART5SMENW

Bit 20 - USART5 clock enable during sleep mode bit

pub fn i2c3smen(&mut self) -> _I2C3SMENW

Bit 30 - I2C3 clock enable during sleep mode bit